moved es _sanitize to its own helper
