+++ usaspending_api/references/v2/views/city.py
@@ -6,7 +6,7 @@ from collections import OrderedDict
 from usaspending_api.common.cache_decorator import cache_response
 
 from usaspending_api.common.elasticsearch.client import es_client_query
-from usaspending_api.search.v2.elasticsearch_helper import es_sanitize
+from usaspending_api.search.v2.es_sanitization import es_sanitize
 from usaspending_api.common.validator.tinyshield import validate_post_request
 from usaspending_api.awards.v2.filters.location_filter_geocode import ALL_FOREIGN_COUNTRIES
