+++ usaspending_api/search/v2/es_sanitization.py
@@ -1,0 +1,14 @@
+import logging
+import re
+
+logger = logging.getLogger("console")
+
+
+def es_sanitize(input_string):
+    """ Escapes reserved elasticsearch characters and removes when necessary """
+
+    processed_string = re.sub(r'([-&!|{}()^~*?:\\/"+\[\]<>])', "", input_string)
+    if len(processed_string) != len(input_string):
+        msg = "Stripped characters from input string New: '{}' Original: '{}'"
+        logger.info(msg.format(processed_string, input_string))
+    return processed_string
